refactor: separate client transport and endpoint responsibilities
